% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namepref0.R
\name{namepref0}
\alias{namepref0}
\title{Add variable name prefix (without substitution)}
\usage{
namepref0(dataframe, pref_new)
}
\arguments{
\item{dataframe}{A dataframe}

\item{pref_new}{New prefix character to be added to existing varnames}
}
\value{
Returns dataframe with those variables renamed, that start with the old prefix.
}
\description{
This function adds the new prefix and changes the dataframe's names to the new ones.
}
\examples{
bla <- tibble::tibble(x_ar = 1:5, y_ar = 6:10)
blo <- namepref0(bla, "z_")
names(bla)
# "x_ar" "y_ar"
names(blo)
# "z_x_ar" "z_y_ar"

}
