% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namesuff.R
\name{namesuff}
\alias{namesuff}
\title{Add variable name suffix (with substitution)}
\usage{
namesuff(dataframe, suf_old, suf_new)
}
\arguments{
\item{dataframe}{A dataframe}

\item{suf_old}{Old suffix character that some or all variables in the dataframe have}

\item{suf_new}{New suffix character to replace the old one}
}
\value{
Returns dataframe with those variables renamed that end with the old suffix.
}
\description{
This function strips the old suffix from the variable names of dataframe.
It then adds the new suffix where the old one was and changes the dataframe's names to the new ones.
}
\examples{
bla <- tibble::tibble(ar_x = 1:5, ar_y = 6:10)
blo <- namesuff(bla, "_x", "_z")
names(bla)
names(blo)

}
