% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namesuff0.R
\name{namesuff0}
\alias{namesuff0}
\title{Add variable name suffix (without substitution)}
\usage{
namesuff0(dataframe, suf_new)
}
\arguments{
\item{dataframe}{A dataframe}

\item{suf_new}{New suffix character to add to the variable names of the dataframe}
}
\value{
Returns dataframe with all variables renamed to end with the new suffix.
}
\description{
This function adds the new suffix and changes
the dataframe's names to the new ones.
}
\examples{
bla <- tibble::tibble(ar_x = 1:5, ar_y = 6:10)
blo <- namesuff(bla, "_z")
names(bla)
names(blo)

}
