% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suffstrip.R
\name{suffstrip}
\alias{suffstrip}
\title{Remove a suffix from all variable names in a datafrme}
\usage{
suffstrip(dataframe, suf_old)
}
\arguments{
\item{dataframe}{A dataframe}

\item{suf_old}{Old suffix to be removed from all variables}
}
\value{
Returns a dataframe with all variables renamed that used to have the old suffix.
Any variable without the old suffix will be kept without changes.
}
\description{
Remove a suffix from all variable names in a datafrme
}
\examples{
bla <- tibble::tibble(ar_x = 1:5, ar_y = 6:10)
blo <- namesuff(bla, "_x")
names(bla)
names(blo)

}
